requirements.txt
streamlit>=1.32
streamlit-drawable-canvas>=0.9.3
gspread>=6.0
google-auth>=2.28
google-api-python-client>=2.120
Pillow>=10.0
numpy>=1.26
